+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/abstracts/extended/ExtendedPrivateChat.kt
@@ -2,4 +2,6 @@ package dev.inmo.tgbotapi.types.chat.abstracts.extended
 
 import dev.inmo.tgbotapi.types.chat.abstracts.PrivateChat
 
-interface ExtendedPrivateChat : PrivateChat, ExtendedChat
+interface ExtendedPrivateChat : PrivateChat, ExtendedChat {
+    val bio: String
+}