+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/extended/ExtendedPrivateChatImpl.kt
@@ -16,5 +16,7 @@ data class ExtendedPrivateChatImpl(
     @SerialName(firstNameField)
     override val firstName: String = "",
     @SerialName(lastNameField)
-    override val lastName: String = ""
+    override val lastName: String = "",
+    @SerialName(bioField)
+    override val bio: String = ""
 ) : ExtendedPrivateChat